requirements.txt
openai-whisper
pyaudio
numpy
torch
torchaudio